Diverse

+++ Stored Procedures/dbo.sp_load_kpi_rookmelder_percentage_aanwezig.sql
@@ -4,13 +4,15 @@ SET ANSI_NULLS ON
 GO
 
 
-create PROCEDURE[dbo].[sp_load_kpi_rookmelder_aantal](
+
+
+CREATE PROCEDURE[dbo].[sp_load_kpi_rookmelder_percentage_aanwezig](
   @peildatum date = '20210930'
 )
 as
 /* #################################################################################################################
 
-exec staedion_dm.[dbo].[sp_load_kpi_rookmelder_aantal] '20210930'
+exec staedion_dm.[dbo].[sp_load_kpi_rookmelder_percentage_aanwezig] '20211005'
 select * from empire_staedion_Data.etl.LogboekMeldingenProcedures
 		declare @fk_indicator_id as smallint
 		select @fk_indicator_id = min(id) from  [Dashboard].[Indicator] where lower([Omschrijving]) like '%personeelslasten%'
@@ -33,9 +35,10 @@ BEGIN TRY
 
 
 -----------------------------------------------------------------------------------------------------------
-	-- 307 Aantal aangebrachte rookmelders
-
-		select @fk_indicator_id = 307
+	-- 308 Percentage woningen met rookmelders
+	IF @peildatum between dateadd(d, 1-day(cast(getdate() as date)), cast(getdate() as date)) and eomonth(cast(getdate() as date))
+		BEGIN
+		select @fk_indicator_id = 308
 		
 			delete from Dashboard.[RealisatieDetails] where fk_indicator_id = @fk_indicator_id and datum between dateadd(d, 1-day(@peildatum), @peildatum) and eomonth(@peildatum);
 
@@ -48,23 +51,22 @@ BEGIN TRY
 				 ),				
 			[Details] as
 				(
-						SELECT   Datum = cast(MEL.[Install_ Date] as date)
-								,Waarde = IIF(MEL.[Install_ Date] is not null, 1, 0)
+						SELECT	 Datum = @peildatum
+							    ,Waarde = IIF(MEL.[Install_ Date] is not null, 1, 0)
 								,Laaddatum = getdate()
-								,Omschrijving = OGE.Nr_ + ' ; ' + 'Merk: ' + IIF(MEL.Brand <> '', MEL.Brand, 'ONBEKEND/NVT') + ' ; ' + 'Type: ' + IIF(MEL.[Type] <> '', MEL.[Type], 'ONBEKEND/NVT')
+								,Omschrijving = OGE.Nr_ + ' ; ' + 'Installatiedatum: ' + IIF(cast(MEL.[Install_ Date] as date) <> cast('17530101' as date), format(cast(MEL.[Install_ Date] as date), 'yyyy-MM-dd'), 'ONBEKEND/NVT') + ' ; ' + 'Merk: ' + IIF(MEL.Brand <> '', MEL.Brand, 'ONBEKEND/NVT') + ' ; ' + 'Type: ' + IIF(MEL.[Type] <> '', MEL.[Type], 'ONBEKEND/NVT')
 								,Eenheidnummer = OGE.Nr_
-								,Clusternummer = MEL.[Technical Cluster No_]
-						FROM empire_data.dbo.[Staedion$OGE] AS OGE
-						left outer join cteFIX AS MEL
-						ON OGE.Nr_ = MEL.No_
-						JOIN empire_Data.dbo.[Staedion$type] AS TT
-						ON TT.[Code] = OGE.[Type] AND TT.Soort = 0
-						WHERE (TT.[Analysis Group Code] = 'WON ZELF' or TT.[Analysis Group Code] = 'WON ONZ')
-						AND OGE.[Begin exploitatie] <> '17530101'
-						AND OGE.[Begin exploitatie] <= @peildatum
-						AND (OGE.[Einde exploitatie] >= @peildatum
-						OR OGE.[Einde exploitatie] = '17530101')
-						AND cast(MEL.[Install_ Date] as date) between dateadd(d, 1-day(@peildatum), @peildatum) and eomonth(@peildatum)
+								,Clusternummer = OGE.[PMC Nr_]
+								FROM empire_data.dbo.[Staedion$OGE] AS OGE
+								left outer join cteFIX AS MEL
+								ON OGE.Nr_ = MEL.No_
+								JOIN empire_Data.dbo.[Staedion$type] AS TT
+								ON TT.[Code] = OGE.[Type] AND TT.Soort = 0
+								WHERE (TT.[Analysis Group Code] = 'WON ZELF' or TT.[Analysis Group Code] = 'WON ONZ')
+								AND OGE.[Begin exploitatie] <> '17530101'
+								AND OGE.[Begin exploitatie] <= @peildatum
+								AND (OGE.[Einde exploitatie] >= @peildatum
+								OR OGE.[Einde exploitatie] = '17530101')
 				)
 
 			insert into [Dashboard].[RealisatieDetails]
@@ -100,11 +102,11 @@ BEGIN TRY
 				Waarde,
 				Laaddatum
 				)
-				select det.fk_indicator_id, @peildatum, sum([Waarde]*1.00), getdate()
+				select det.fk_indicator_id, @peildatum, avg([Waarde]*1.00), getdate()
 				from Dashboard.[RealisatieDetails] det
 				where det.fk_indicator_id = @fk_indicator_id and det.Datum between dateadd(d, 1-day(@peildatum), @peildatum) and @peildatum
 				group by det.fk_indicator_id;
-
+		END
 -----------------------------------------------------------------------------------------------------------
 
 		set		@finish = current_timestamp
